Refactor GetRestConfig as an apiserver function
