--- pkg/services/unifiedstorage/unifiedstorage.go
@@ -1,10 +1,0 @@
-package unifiedstorage
-
-import (
-	"github.com/grafana/grafana/pkg/services/apiserver"
-)
-
-type DependencyRegistry interface {
-	RegisterRestConfigProvider(restConfigProvider apiserver.RestConfigProvider)
-	GetRestConfigProvider() apiserver.RestConfigProvider
-}
